\name{transmartRClient-package}
\alias{transmartRClient-package}
\alias{transmartRClient}
\docType{package}
\title{
R Client for accessing the tranSMART RESTful API
}
\description{
Contains functions that enable a user to authenticate and access
a tranSMART database via get-functions which access tranSMART's
RESTful API.
}
\details{
  \tabular{ll}{
    Package: \tab transmartRClient\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2014-02-25\cr
    License: \tab \cr
  }
}

\author{
  The Hyve
  Maintainer: tim@thehyve.nl
}
\references{
}
\keyword{ package }
