\name{transmartRClient-package}
\alias{transmartRClient-package}
\alias{transmartRClient}
\docType{package}
\title{
R Client for accessing the tranSMART RESTful API
}
\description{
Contains functions that enable a user to authenticate and access
a tranSMART database via get-functions which access tranSMART's
RESTful API. It currently allows a user to retrieve both clinical
and high dimensional data (to which the user has access).
}
\details{
  \tabular{ll}{
    Package: \tab transmartRClient\cr
    Type: \tab Package\cr
    Version: \tab 0.2\cr
    Date: \tab 2014-02-25\cr
    License: \tab GPL-3 and higher \cr
  }
}

\author{Tim Dorscheidt, Jan Kanis, Rianne Jansen. 
Contact: development@thehyve.nl}
\references{
}
\keyword{ package }
